Mostly Integrated Into Recording Creation Screen
It's ugly but the basics are there. The logic definitely needs some cleanup, but if you enter a title and press photo, it will let you take a photo which will be saved to memory and then replace the little icon when you come back. The  photo is in persistent memory, but will get deleted when you uninstall the app